elf: Move symbol table code to separate module
Move all structs and code handling the ELF symbol table to a new
submodule.
